+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+- 2026-02-24: configure long-running queries to avoid errors.
 - 2026-02-23:
   - updated packages.
   - removed legacy stub for Blazor demo.